+++ Cargo.toml
@@ -18,3 +18,4 @@ raii-counter = "0.1"
 [dev-dependencies]
 env_logger = "0.5"
 lazy_static = "1.0"
+ipnet = "1.2"
