+++ src/config.rs
@@ -18,16 +18,3 @@ pub struct Config {
     /// Number of DNS threads per worker
     pub dns_threads_per_worker: usize,
 }
-
-#[cfg(test)]
-impl Default for Config {
-    fn default() -> Config {
-        Config {
-            keep_alive_timeout: Duration::from_secs(3),
-            transaction_timeout: Duration::from_secs(10),
-            max_transactions_per_worker: 1_000,
-            workers: 2,
-            dns_threads_per_worker: 10,
-        }
-    }
-}
